App_backend/config.py
import os
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

MONGO_URI = os.getenv("MONGO_URI")
SECRET_KEY = os.getenv("SECRET_KEY")
JWT_SECRET_KEY= os.getenv("JWT_SECRET_KEY")
